requirements.txt
Django==3.2.9
mysql-connector
mysqlclient
configparser
ldap3
django_auth_ldap
pyldap
python-environ
django-rosetta
django-parler
django-crispy-forms
django-flatpickr
django-bootstrap-datepicker-plus
django-bootstrap4
django-readonly-field
django-tables2
django-filter
django-widget-tweaks
django-utils-six 
django-import-export